+++ jarvis/api/squire/scheduler.py
@@ -2,7 +2,7 @@ import os
 import shutil
 
 from jarvis.api import triggers
-from jarvis.modules.crontab.expression import CronExpression
+from jarvis.modules.crontab import expression
 from jarvis.modules.utils import util
 
 
@@ -33,7 +33,7 @@ class MarketHours:
     }
 
 
-def rh_cron_schedule(extended: bool = False) -> CronExpression:
+def rh_cron_schedule(extended: bool = False) -> expression.CronExpression:
     """Creates a cron expression for ``stock_report.py``. Determines cron schedule based on current timezone.
 
     Args:
@@ -53,10 +53,10 @@ def rh_cron_schedule(extended: bool = False) -> CronExpression:
         tz = 'OTHER'
     start = MarketHours.hours['EXTENDED'][tz]['OPEN'] if extended else MarketHours.hours['REGULAR'][tz]['OPEN']
     end = MarketHours.hours['EXTENDED'][tz]['CLOSE'] if extended else MarketHours.hours['REGULAR'][tz]['CLOSE']
-    return CronExpression(f"*/30 {start}-{end} * * 1-5 {job}")
+    return expression.CronExpression(f"*/30 {start}-{end} * * 1-5 {job}")
 
 
-def sm_cron_schedule(include_weekends: bool = True) -> CronExpression:
+def sm_cron_schedule(include_weekends: bool = True) -> expression.CronExpression:
     """Creates a cron expression for ``stock_monitor.py``.
 
     Args:
@@ -68,5 +68,5 @@ def sm_cron_schedule(include_weekends: bool = True) -> CronExpression:
     """
     job = f"cd {os.getcwd()} && {shutil.which(cmd='python')} {os.path.join(triggers.__path__[0], 'stock_monitor.py')}"
     if include_weekends:
-        return CronExpression(f"*/15 * * * * {job}")
-    return CronExpression(f"*/15 * * * 1-5 {job}")
+        return expression.CronExpression(f"*/15 * * * * {job}")
+    return expression.CronExpression(f"*/15 * * * 1-5 {job}")