+++ jarvis/api/squire/stockanalysis_squire.py
@@ -10,7 +10,7 @@ import yaml
 from bs4 import BeautifulSoup
 from webull import webull
 
-from jarvis.api.modals import settings
+from jarvis.api.models import settings
 from jarvis.api.squire.logger import logger
 from jarvis.modules.exceptions import EgressErrors
 from jarvis.modules.models import models