+++ jarvis/api/squire/timeout_otp.py
@@ -2,13 +2,12 @@ from typing import NoReturn
 
 from pydantic import EmailStr
 
-from jarvis.api.modals.settings import (robinhood, stock_monitor_helper,
-                                        surveillance)
+from jarvis.api.models import settings
 
 
 def reset_robinhood() -> NoReturn:
     """Resets robinhood token after the set time."""
-    robinhood.token = None
+    settings.robinhood.token = None
 
 
 def reset_stock_monitor(email_address: EmailStr) -> NoReturn:
@@ -17,10 +16,10 @@ def reset_stock_monitor(email_address: EmailStr) -> NoReturn:
     Args:
         email_address: Email address that should be cleared.
     """
-    if stock_monitor_helper.otp_sent.get(email_address):
-        del stock_monitor_helper.otp_sent[email_address]
+    if settings.stock_monitor_helper.otp_sent.get(email_address):
+        del settings.stock_monitor_helper.otp_sent[email_address]
 
 
 def reset_surveillance() -> NoReturn:
     """Resets surveillance token after the set time."""
-    surveillance.token = None
+    settings.surveillance.token = None