+++ jarvis/modules/logger/config.py
@@ -12,8 +12,7 @@ from logging import Formatter
 from pydantic import BaseModel
 
 from jarvis.modules.builtin_overrides import AddProcessName
-from jarvis.modules.logger.custom_logger import (DEFAULT_LOG_FORM,
-                                                 custom_handler, logger)
+from jarvis.modules.logger import custom_logger
 from jarvis.modules.models import models
 
 
@@ -31,16 +30,16 @@ def multiprocessing_logger(filename: str, log_format: Formatter = None) -> str:
         str:
         Actual log filename with datetime converted.
     """
-    logger.propagate = False
+    custom_logger.logger.propagate = False
     # Remove existing handlers
-    for _handler in logger.handlers:
-        logger.removeHandler(hdlr=_handler)
-    log_handler = custom_handler(filename=filename, log_format=log_format)
-    logger.addHandler(hdlr=log_handler)
+    for _handler in custom_logger.logger.handlers:
+        custom_logger.logger.removeHandler(hdlr=_handler)
+    log_handler = custom_logger.custom_handler(filename=filename, log_format=log_format)
+    custom_logger.logger.addHandler(hdlr=log_handler)
     # Remove existing filters from the new log handler
-    for _filter in logger.filters:
-        logger.removeFilter(_filter)
-    logger.addFilter(filter=AddProcessName(process_name=models.settings.pname))
+    for _filter in custom_logger.logger.filters:
+        custom_logger.logger.removeFilter(_filter)
+    custom_logger.logger.addFilter(filter=AddProcessName(process_name=models.settings.pname))
     return log_handler.baseFilename
 
 
@@ -56,7 +55,7 @@ class APIConfig(BaseModel):
     ACCESS_LOG_FILENAME = datetime.now().strftime(os.path.join('logs', 'fast_api_access_%d-%m-%Y.log'))
     DEFAULT_LOG_FILENAME = datetime.now().strftime(os.path.join('logs', 'fast_api_%d-%m-%Y.log'))
 
-    DEFAULT_LOG_FORMAT = DEFAULT_LOG_FORM
+    DEFAULT_LOG_FORMAT = custom_logger.DEFAULT_LOG_FORM
     ACCESS_LOG_FORMAT = '%(levelprefix)s %(client_addr)s - "%(request_line)s" %(status_code)s'
     ERROR_LOG_FORMAT = '%(levelname)s\t %(message)s'